Extract factory method for SUT creation
I can't move in small steps as long as the constructor sets the SUT as a
class field. This would require me to replace each of the dependencies
wholesale, instead of one test at a time.

To be explicit: There are only three tests, so I could easily do this
within a Pomodoro (25 minutes), but for those who are interested in the
process in addition to the final result, I'll demonstrate how to move in
smaller steps.