Inline CreateController in one test
Now the test is directly calling the `Controller` constructor. Isn't
this a maintenance problem? What if you want to change the constructor
signature?

Indeed, this can be a maintenance problem if you use a lot of DI.
There's a few way to address such an issue:
- Create a proper Immutable Fluent Builder for the class, if need be.
  Just like a Test Data Builder, but for the SUT instead of test data.
- Keep going with replacing polymorphic dependencies with pure
  functions. As future commits will show, the `Controller` constructor
  will turn out to be simpler in the end.
You can combine the above two options, if need be. They aren't mutually
exclusive.