Remove code as an explicit test parameter
The previous state of the test was actually a lie. The impliciation was
that `code` could be any value, entirely disconnected from `knownState`,
where the reality is that these two values must match.

If you regard tests as executable specification, the previous test was a
poor specification, because it was being very implicit about what
'valid' actually means.

The two new lines of code in line 27 and 28 are perhaps a bit verbose,
but it should make the relationship more clean than if I had just
inlined `knownState.Item1`, I hope.