Add guard clause to StateValidator
This is really not necessary, as a pure validator could easily just
return `false` in this case - which I believe that it already did.

From correspondence with Sergei I got the impression, however, that
handling an exception case was an important part of the example, so I
don't want to simplify the `Controller` behaviour, even though the type
system quite clearly makes it impossible for the original
`StateValidator` to throw an exception... or, at least, I couldn't
figure out how to provoke one in about fifteen minutes of trial and
error.