Refactor test to use pure StateValidator
As I hinted at in the previous commit message, I consider this test
essentially superfluous, since the only implementation of
`IStateValidator` couldn't (as far as I could tell) throw exceptions.
Still, to go with the spirit of the orginal example, I've kept it by
adding some exception-throwing behaviour to `StateValidator`.