Use pure Renderer in test case
This test case is, again, a bit iffy. I can get by with expecting "500"
because the default `isMobile` value is `false`. However, one could
argue that it'd be more appropriate to create an `expected` value by
calling `renderer.Error`, equivalent to the two previous test cases.
This, however, would require that the test passes exectly the same
`ArgumentNullException` to `renderer.Error` that `StateValidator` would
create. This would couple the test to an implementation detail, which is
always a tension with this more composed (dare I say 'integration')
tests.

Usually, I'd rely on the unit testing of `Renderer` and be a little more
relaxed in my assertions on the `Controller` level.

This does, however, also highlight how, even when working with pure
functions, it's still a good idea to listen to the tests. This test is
problematic, which may imply one of two things:
- The test is poorly written
- The SUT is inadequately designed
I'd lean towards the latter interpretation and blame the whole exception
mechanism rather than the test. Using Eithers/Results tends to give you
better test affordances as well as all the other benefits.